Use hardcoded version as the default
GH does only have the concept of "latest" version which is the latest from time point of view.
In a shell script we cannot really do semver comparison and thus we cannot determine which is the greatest from semver point of view and thus have to rely on hardcoded value.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>